+++ tests/index.js
@@ -4,7 +4,7 @@ const readFileSync = require( 'fs' ).readFileSync;
 const chai = require( 'chai' );
 const expect = chai.expect;
 const rollup = require( 'rollup' ).rollup;
-const plugin = require( '../dist/rollup-plugin-babili' );
+const plugin = require( '../dist/rollup-plugin-babel-minify' );
 const babel = require( 'babel-core' );
 const validateSourcemap = require( 'sourcemap-validator' );
 
@@ -14,15 +14,15 @@ const bundleOptions = {
 
 process.chdir( 'tests' );
 
-describe( 'rollup-plugin-babili', () => {
+describe( 'rollup-plugin-babel-minify', () => {
 	it( 'is a function', () => {
 		expect( plugin ).to.be.a( 'function' );
 	} );
 
-	it( 'minifies code just like Babili', () => {
+	it( 'minifies code just like babel-minify', () => {
 		const path = 'fixtures/index.js';
 		const code = readFileSync( path, 'utf8' );
-		const babeledCode = babel.transform( code, { presets: [ 'babili' ], comments: true } );
+		const babeledCode = babel.transform( code, { presets: [ 'minify' ], comments: true } );
 
 		return rollup( {
 			entry: path,
@@ -37,7 +37,7 @@ describe( 'rollup-plugin-babili', () => {
 	it( 'passes options to babel', () => {
 		const path = 'fixtures/sourcemap.js';
 		const code = readFileSync( path );
-		const babeledCode = babel.transform( code, { presets: [ [ 'babili', {
+		const babeledCode = babel.transform( code, { presets: [ [ 'minify', {
 			removeConsole: true
 		} ] ] } );
 
@@ -56,7 +56,7 @@ describe( 'rollup-plugin-babili', () => {
 	it( 'removes comments', () => {
 		const path = 'fixtures/index.js';
 		const code = readFileSync( path );
-		const babeledCode = babel.transform( code, { presets: [ 'babili' ], comments: false } );
+		const babeledCode = babel.transform( code, { presets: [ 'minify' ], comments: false } );
 
 		return rollup( {
 			entry: path,