+++ CHANGELOG.md
@@ -1,10 +1,14 @@
-# rollup-plugin-babili Changelog
+# rollup-plugin-babel-minify Changelog
 
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ---
+
+## 3.1.2
+### Changed
+* Change name from `rollup-plugin-babili` to `rollup-plugin-babel-minify` and update dependencies accordingly.
 
 ## [3.1.1] – 2017-08-12
 ### Changed
@@ -59,6 +63,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * First working version, yay!
 
+[3.1.1]: https://github.com/Comandeer/rollup-plugin-babili/compare/v3.1.0...v3.1.1
 [3.1.0]: https://github.com/Comandeer/rollup-plugin-babili/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/Comandeer/rollup-plugin-babili/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/Comandeer/rollup-plugin-babili/compare/v1.1.1...v2.0.0